+++ src/dropin.js
@@ -161,10 +161,12 @@ Dropin.prototype._initialize = function (callback) {
 
   // Backfill with `en`
   strings = assign({}, translations.en);
-  if (this._merchantConfiguration.locale) {
+  if (typeof this._merchantConfiguration.locale === 'string') {
     localizedStrings = translations[this._merchantConfiguration.locale] || translations[this._merchantConfiguration.locale.split('_')[0]];
     // Fill `strings` with `localizedStrings` that may exist
     strings = assign(strings, localizedStrings);
+  } else if (typeof this._merchantConfiguration.locale === 'object') {
+    strings = assign(strings, this._merchantConfiguration.locale);
   }
 
   localizedHTML = Object.keys(strings).reduce(function (result, stringKey) {