+++ src/views/payment-sheet-views/base-paypal-view.js
@@ -32,10 +32,9 @@ BasePayPalView.prototype._initialize = function (isCredit) {
   }, ASYNC_DEPENDENCY_TIMEOUT);
 
   btPaypal.create({client: this.client}, function (err, paypalInstance) {
-    var checkoutJSConfiguration;
+    var checkoutJSConfiguration, locale;
     var buttonSelector = '[data-braintree-id="paypal-button"]';
     var environment = self.client.getConfiguration().gatewayConfiguration.environment === 'production' ? 'production' : 'sandbox';
-    var locale = self.model.merchantConfiguration.locale;
 
     if (err) {
       self.model.asyncDependencyFailed({
@@ -43,6 +42,10 @@ BasePayPalView.prototype._initialize = function (isCredit) {
         error: err
       });
       return;
+    }
+
+    if (typeof self.model.merchantConfiguration.locale === 'string') {
+      locale = self.model.merchantConfiguration.locale;
     }
 
     self.paypalInstance = paypalInstance;